b3-trading-platform/services/market-data/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0